+++ CHANGELOG.md
@@ -25,8 +25,7 @@ All notable changes to this project will be documented here.
 - menu_parse() - optimized processing
 - Added run_if_found() - makes code cleaner
 - Bug fix: fix_home_path() - Now correctly handles the odd behaviour of tmux 3.4
-
----
+- moved all profiling init to dbg_profiling.sh
 
 ## [2.2.3] - 2025-04-28